Added an abstract class for Broadcasters.
